Remove revm from workspace
