+++ Cargo.toml
@@ -21,10 +21,8 @@ members = [
   "crates/storage",
   "crates/vm",
   "crates/vm/levm",
-  "crates/vm/levm/bench/revm_comparison",
   "crates/vm/levm/runner",
   "tooling/ef_tests/blockchain",
-  "tooling/ef_tests/state",
   "tooling/ef_tests/state_v2",
   "tooling/genesis",
   "tooling/hive_report",
@@ -101,7 +99,7 @@ rand = "0.8.5"
 cfg-if = "1.0.0"
 reqwest = { version = "0.12.7", features = ["socks", "json"] }
 snap = "1.1.1"
-secp256k1 = { version = "0.29.1", default-features = false, features = [
+secp256k1 = { version = "0.30.0", default-features = false, features = [
   "global-context",
   "recovery",
   "rand",
@@ -128,4 +126,4 @@ tower-http = { version = "0.6.2", features = ["cors"] }
 rocksdb = "0.24.0"
 
 [patch.crates-io]
-secp256k1 = { git = "https://github.com/sp1-patches/rust-secp256k1", tag = "patch-0.29.1-sp1-5.0.0" }
+secp256k1 = { git = "https://github.com/sp1-patches/rust-secp256k1", tag = "patch-0.30.0-sp1-5.0.0" }
