+++ cmd/ethrex/initializers.rs
@@ -119,9 +119,6 @@ pub fn init_blockchain(
     blockchain_type: BlockchainType,
     perf_logs_enabled: bool,
 ) -> Arc<Blockchain> {
-    #[cfg(feature = "revm")]
-    info!("Initiating blockchain with revm");
-    #[cfg(not(feature = "revm"))]
     info!("Initiating blockchain with levm");
     Blockchain::new(store, blockchain_type, perf_logs_enabled).into()
 }
