+++ tooling/ef_tests/state/runner/mod.rs
@@ -1,7 +1,6 @@
 use std::collections::BTreeMap;
 
 use crate::report::EFTestsReport;
-#[cfg(feature = "revm")]
 use crate::runner::revm_runner::SpecId;
 use crate::{
     parser::SPECIFIC_IGNORED_TESTS,
@@ -18,6 +17,7 @@ use spinoff::{Color, Spinner, spinners::Dots};
 
 pub mod levm_runner;
 pub mod revm_runner;
+pub mod revm_db;
 
 #[derive(Debug, thiserror::Error, Clone, Serialize, Deserialize)]
 pub enum EFTestRunnerError {
